Refactor ast utils
